add llm name to the prompt
