+++ tests/res/optimize/llm_cache_gpt-3.5-turbo_9a4540619ef903566aa0a958d20e8463.jsonl
@@ -1,3 +1,3 @@
-["a61524f40181a79f34bcf3bb2bfab0d6", {"chunk": null, "output": {"content": " produce the query. We know that David Gregory inherited a castle, and we are curious about the number of storeys it has. To find this information, we can search for the architectural details of the castle he inherited.\nQuery: \"architectural details of castle inherited by David Gregory\"", "role": "assistant"}}]
-["3e9b4479fdf62f58542ebf074b3d67af", {"chunk": null, "output": {"content": " produce the query. We know that David Gregory inherited Kinnairdy Castle in 1664. To find out how many storeys are in the castle, we need to search for information about the castle's structure.\nQuery: \"Kinnairdy Castle storeys\"", "role": "assistant"}}]
-["28c321c5a0c14e2b0f1889c079bf8758", {"chunk": null, "output": {"content": "produce the answer. We know that Kinnairdy Castle, which David Gregory inherited, has five storeys and a garret.\nAnswer: Five", "role": "assistant"}}]
+["74e9bac1ba9bf8ccfeebeabe8f9f7df9", {"chunk": null, "output": {"content": " produce the query. We know that David Gregory inherited a castle, and we are curious about the number of storeys it has. To find this information, we can search for the architectural details of the castle he inherited.\nQuery: \"Architectural details of castle inherited by David Gregory\"", "role": "assistant", "tool_calls": null, "function_call": null}}]
+["1f1f5c0c64314fbdb29004cbd3c59603", {"chunk": null, "output": {"content": " produce the query. We know that David Gregory inherited Kinnairdy Castle in 1664. To find out how many storeys are in the castle, we need to search for information about the castle's structure.\nQuery: \"Kinnairdy Castle storeys\"", "role": "assistant", "tool_calls": null, "function_call": null}}]
+["6ef5b53f2624adb51a38a85a1b3c3315", {"chunk": null, "output": {"content": "Answer: five", "role": "assistant", "tool_calls": null, "function_call": null}}]
