+++ tests/test_examples.py
@@ -15,6 +15,7 @@ from omegaconf import DictConfig
 from examples.gsm8k_tuning.finetune_student import get_training_samples_from_tapes
 from tapeagents.finetune.data import load_samples
 from tapeagents.io import load_tapes
+from tests.test_utils import load_tape_dict, load_traces
 
 sys.path.append(str(Path(__file__).parent.parent.resolve()))  # allow to import from examples
 
@@ -74,20 +75,6 @@ def set_sqlite_db_dir(dir: str):
             os.environ["TAPEAGENTS_SQLITE_DB"] = old_path
         else:
             del os.environ["TAPEAGENTS_SQLITE_DB"]
-
-
-def load_tape_dict(run_dir: str, fname: str = "tape.json") -> dict:
-    tape_fpath = os.path.join(run_dir, fname)
-    with open(tape_fpath, "r") as f:
-        tape_dict = json.load(f)
-    return tape_dict
-
-
-def load_traces(run_dir: str, fname: str = "traces.json") -> list[TrainingText]:
-    traces_fpath = os.path.join(run_dir, fname)
-    with open(traces_fpath, "r") as f:
-        traces = [TrainingText.model_validate(t) for t in json.load(f)]
-    return traces
 
 
 def test_llama_agent():
